Move hive configs to common location within hudi-sync common
