+++ docker/demo/config/test-suite/templates/test.properties.template
@@ -46,5 +46,5 @@ hoodie.datasource.hive_sync.database=testdb
 hoodie.datasource.hive_sync.table=table1
 hoodie.datasource.hive_sync.assume_date_partitioning=false
 hoodie.datasource.hive_sync.partition_fields=_hoodie_partition_path
-hoodie.datasource.hive_sync.partition_extractor_class=org.apache.hudi.hive.SlashEncodedDayPartitionValueExtractor
+hoodie.datasource.hive_sync.partition_extractor_class=org.apache.hudi.sync.common.SlashEncodedDayPartitionValueExtractor
