+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/hudi/TestDataSourceOptions.scala
@@ -18,8 +18,8 @@
 package org.apache.hudi
 
 import org.apache.hudi.DataSourceWriteOptions._
-import org.apache.hudi.hive.{HiveStylePartitionValueExtractor, MultiPartKeysValueExtractor}
 import org.apache.hudi.keygen.{ComplexKeyGenerator, SimpleKeyGenerator}
+import org.apache.hudi.sync.common.{HiveStylePartitionValueExtractor, MultiPartKeysValueExtractor}
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
