+++ hudi-sync/hudi-hive-sync/src/test/java/org/apache/hudi/hive/TestMultiPartKeysValueExtractor.java
@@ -18,6 +18,8 @@
 
 package org.apache.hudi.hive;
 
+import org.apache.hudi.sync.common.MultiPartKeysValueExtractor;
+
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.Test;
