+++ hudi-sync/hudi-sync-common/src/main/java/org/apache/hudi/sync/common/MultiPartKeysValueExtractor.java
@@ -16,27 +16,36 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.hive;
+package org.apache.hudi.sync.common;
 
 import java.util.Collections;
+import org.apache.hudi.common.util.ValidationUtils;
+import org.apache.hudi.sync.common.PartitionValueExtractor;
+
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
- * Extractor for Hive Style Partitioned tables, when the parition folders are key value pairs.
- *
- * <p>This implementation extracts the partition value of yyyy-mm-dd from the path of type datestr=yyyy-mm-dd.
+ * Partition Key extractor treating each value delimited by slash as separate key.
  */
-public class HiveStylePartitionValueExtractor implements PartitionValueExtractor {
-  private static final long serialVersionUID = 1L;
+public class MultiPartKeysValueExtractor implements PartitionValueExtractor {
 
   @Override
   public List<String> extractPartitionValuesInPath(String partitionPath) {
-    // partition path is expected to be in this format partition_key=partition_value.
-    String[] splits = partitionPath.split("=");
-    if (splits.length != 2) {
-      throw new IllegalArgumentException(
-              "Partition path " + partitionPath + " is not in the form partition_key=partition_value.");
+    // If the partitionPath is empty string( which means none-partition table), the partition values
+    // should be empty list.
+    if (partitionPath.isEmpty()) {
+      return Collections.emptyList();
     }
-    return Collections.singletonList(splits[1]);
+    String[] splits = partitionPath.split("/");
+    return Arrays.stream(splits).map(s -> {
+      if (s.contains("=")) {
+        String[] moreSplit = s.split("=");
+        ValidationUtils.checkArgument(moreSplit.length == 2, "Partition Field (" + s + ") not in expected format");
+        return moreSplit[1];
+      }
+      return s;
+    }).collect(Collectors.toList());
   }
 }
