+++ hudi-sync/hudi-sync-common/src/main/java/org/apache/hudi/sync/common/PartitionValueExtractor.java
@@ -16,18 +16,18 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.hive;
+package org.apache.hudi.sync.common;
 
-import java.util.ArrayList;
+import java.io.Serializable;
 import java.util.List;
 
 /**
- * Extractor for Non-partitioned hive tables.
+ * HDFS Path contain hive partition values for the keys it is partitioned on. This mapping is not straight forward and
+ * requires a pluggable implementation to extract the partition value from HDFS path.
+ * <p>
+ * e.g. Hive table partitioned by datestr=yyyy-mm-dd and hdfs path /app/hoodie/dataset1/YYYY=[yyyy]/MM=[mm]/DD=[dd]
  */
-public class NonPartitionedExtractor implements PartitionValueExtractor {
+public interface PartitionValueExtractor extends Serializable {
 
-  @Override
-  public List<String> extractPartitionValuesInPath(String partitionPath) {
-    return new ArrayList<>();
-  }
+  List<String> extractPartitionValuesInPath(String partitionPath);
 }
